[MOD#12] home에서 mypage 이동 삭제 및 community와 viewingparty에서 이동하도록 추가
