+++ app/src/main/res/layout/fragment_home.xml
@@ -20,17 +20,5 @@
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintTop_toTopOf="parent"/>
 
-        <!--마이 페이지 대충 구현 -->
-        <Button
-            android:id="@+id/btnmypage"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="My Page"
-            style="@style/TextAppearance.LCK.Light.Body_14sp"
-            android:textColor="@color/white"
-            app:layout_constraintStart_toStartOf="parent"
-            android:layout_marginStart="75dp"
-            app:layout_constraintTop_toTopOf="parent"/>
-
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
