+++ app/src/main/res/layout/fragment_mypage_viewing_party_guest.xml
@@ -18,5 +18,7 @@
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:padding="16dp"/>
+
     </androidx.constraintlayout.widget.ConstraintLayout>
+
 </layout>
